% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof-test.R
\name{gof.test}
\alias{gof.test}
\title{Goodness of fit test}
\usage{
gof.test(
  kiri.first,
  kiri.last,
  kanan.first,
  kanan.last,
  j,
  f.obs,
  var,
  miu,
  par,
  alp
)
}
\arguments{
\item{kiri.first}{Batas bawah kelompok pertama (bukan tepi)}

\item{kiri.last}{Batas bawah kelompok terakhir (bukan tepi)}

\item{kanan.first}{Batas atas kelompok pertama (bukan tepi)}

\item{kanan.last}{Batas atas kelompok terakhir (bukan tepi)}

\item{j}{Jarak antar interval. misal: (140-144) (144-145), maka j = 5}

\item{f.obs}{Frekuensi observasi per kelas}

\item{var}{Varians populasi}

\item{miu}{Rata-rata populasi}

\item{par}{Banyaknya parameter dalam distribusi yang akan dicocokkan}

\item{alp}{Taraf signifikansi yang digunakan}
}
\value{
Menentukan kecocokan distribusi suatu data disertai perhitungan
}
\description{
Fungsi digunakan untuk menentukan apakah suatu data observasi memiliki
kecocokan dengan suatu distribusi.
}
