% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_interval.R
\name{interval.mean}
\alias{interval.mean}
\title{Pendugaan selang rata-rata satu populasi}
\usage{
interval.mean(alp = 0.05, vek = c(10, 10, 10, 10, 10, 10, 10), sig = NA)
}
\arguments{
\item{alp}{Taraf signifikansi}

\item{vek}{Data sampel numeric}

\item{sig}{Nilai standar deviasi dari populasi}
}
\value{
Plot penduga interval menggunakan ggplot2 dari tidyverse
}
\description{
Fungsi digunakan untuk menghitung berbagai penduga interval rata-rata dari populasi.
}
