% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ujiproporsi.R
\name{proportion.test}
\alias{proportion.test}
\title{Pengujian hipotesis perbandingan dua proporsi}
\usage{
proportion.test(
  p0 = 0,
  alp = 0.05,
  h1 = "two.sided",
  x1 = NA,
  n1 = NA,
  x2 = NA,
  n2 = NA
)
}
\arguments{
\item{p0}{Nilai selisih proporsi dua sampel}

\item{alp}{Taraf signifikansi}

\item{h1}{Jenis uji: two.sided -default- ; right.sided ; left.sided}

\item{x1}{Banyak karakteristik pada sampel pertama}

\item{n1}{Banyak sampel pertama}

\item{x2}{Banyak karakteristik pada sampel kedua}

\item{n2}{Banyak sampel kedua}
}
\value{
Plot uji proporsi menggunakan ggplot2
}
\description{
Fungsi digunakan untuk menentukan keputusan dari suatu hipotesis.
Dalam fungsi ini digunakan beberapa parameter yang diperlukan dalam perbandingan
dua proporsi.
}
