% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance-test.R
\name{variance.test}
\alias{variance.test}
\title{Uji Varians 1 atau 2 Populasi}
\usage{
variance.test(
  varsampel = c(NA, NA),
  nsampel = c(NA, NA),
  varpop0 = NA,
  h1 = "two.sided",
  alpha = 0.05
)
}
\arguments{
\item{varsampel}{varians dari sampel (untuk 1 populasi langsung input nilai, untuk 2 populasi gunakan syntax c(), contoh: c(varsampel1, varsampel2))}

\item{nsampel}{jumlah sampel (untuk 1 populasi langsung input nilai, untuk 2 populasi gunakan syntax c(), contoh: c(nsampel1, nsampel2))}

\item{varpop0}{input varians populasi uji untuk uji 1 sampel, apabila pada parameter termuat, maka akan menghasilkan uji varians 1 populasi}

\item{h1}{hipotesis alternatif (HA/H1), pilih = "two.sided", "right.sided", atau "left.sided". default = "two.sided"}

\item{alpha}{taraf signifikansi yang diinginkan}
}
\value{
Uji varians
}
\description{
Fungsi digunakan untuk menguji varians baik dari satu ataupun dua populasi
}
